Support alternative builds
This allows you to specify alternative builds in one manifest file.
You specify the alternative on the commandline like:
  flatpak-builder --alt=NAME ...

And this then builds the manifest like normally, but with a different
cache base-name (so that both alternatives can be cached at the
same time). However, using the new "alt" option in BuildOption
and [only|skip]-alts options in modules and sources you can
affect what gets built and how. This is very similar to how
things can already be per-arch optional.