Ignore the `--metadata` argument of flatpak build-finish
An app may use `--metadata` in the `finish-args` section to
add section/key/value entries in the metadata file.

If we pass this through to `flatpak build`, it's interpreted to
be a filename. Since it's not a filename, this leads to the
metadata file not being found, and an error is raised like this:

    /home/sam/src/gnome-photos/.flatpak-builder/rofiles/rofiles-qtBvVQ not initialized, use flatpak build-init