context: Make FUSE 2 optimizations opt-in
FUSE 3 does not support the splice_write,splice_move options: those
and similar speedups are enabled by default with FUSE 3.

Distributions wanting faster FUSE 2 filesystems can enable this by
configuring --with-fuse=2, but their binaries will not work when
libostree has been recompiled against FUSE 3. This behaviour was
unconditional in older releases such as 1.2.1.

Conversely, we can add opt-in features that require FUSE 3 by
configuring --with-fuse=3 if we want to, although at the moment no such
features exist.

The new default is to be compatible with all libostree versions,
regardless of whether they were compiled for FUSE 2 or 3.

Resolves: https://github.com/flatpak/flatpak-builder/issues/448
Signed-off-by: Simon McVittie <smcv@collabora.com>