ask for default branch name instead of harcoding 'master'
Many projects are starting to change their default branch
name from 'master' to 'main' and similar, eg.:
https://gitlab.gnome.org/GNOME/evince/-/issues/1635

So when a project is not providing a specific branch,
instead of defaulting to 'master' we should ask the
remote repo which is their default branch name.

That can be retrieve with following command:
git ls-remote --symref origin HEAD

where 'origin' can also be a URL to the repository.

Fixes issue #375