build: Fix Meson build when yaml is disabled
This had two issues in the Meson build system: it was always set to true
even if configured with `-Dyaml=disabled`, and the
`#ifdef FLATPAK_BUILDER_ENABLE_YAML` check always succeeded because
`set10` defines the macro to 1 or 0. Use `set` with a boolean value
instead, which defines or undefines the macro, leading to the same
interface as in Autotools.

Signed-off-by: Simon McVittie <smcv@collabora.com>