builder-context: Write a gitignore file into the state-dir
Unless specified, state-dir is placed on the current path,
which is usually the source directory of the project.

Write a .gitignore file to make sure that both the state-dir
never gets checked out into git and that all the tooling
will ignore it and avoid indexing it.

Inpired by meson's gitignore file in the build directories.

Retro-actively discovered the same change in flatpak-build init [1]

[1]: https://github.com/flatpak/flatpak/commit/a62f8cfb62c4232ef2674cfa4431169a292120d6