Enable authenticated downloads via netrc file
Sometimes source archives are only accessible to logged-in users,
for example as described in issue #456. However, authentication
currently is not supported by flatpak-builder aside from hard-coding
credentials in the manifest file.

Since flatpak-builder uses libcurl to download sources we can make
use of cURL's support for netrc files to provide credentials
out-of-band. That way the manifest file can be shared with others
without leaking credentials.

See the cURL documentation [0] for details about where cURL is
looking for the .netrc file and the expected format of that file.

[0] https://curl.se/libcurl/c/CURLOPT_NETRC.html