builder-main: Commit mirrored media as screenshot ref to exported repo
This is needed to store the application icons and the mirrored
screenshots (as done by appstream) to the ostree repo. This can be
useful in many ways, the primary being flat-manager extracts [1] these
screenshot refs and stores them in the server so that application
store frontends can use their own mirrored screenshot and icon URLs.

This has been for the longest time done at various downstream levels
such as in buildbot [2], flatpak-github-actions [3] and vorarbeiter [4]
(the successor to buildbot) but it's more user friendly to have
flatpak-builder do this extra work.

[1]: https://github.com/flatpak/flat-manager/commit/eff43004756c011b0a0d034a61cec84836b8e0a9
[2]: https://github.com/flathub-infra/buildbot/blob/9d8e2b5483b2e446389548446ebf9c3934380682/master/buildbot/flathub_master.py#L1081
[3]: https://github.com/flatpak/flatpak-github-actions/blob/6684584b07d86113b1acfc80b3e4667f16f617c3/flatpak-builder/index.js#L280-L286
[4]: https://github.com/flathub-infra/vorarbeiter/blob/7f2f04a562e73ba4f46044dae7c99f8e9e64a39d/justfile#L185-L189