builder-manifest: Run appstream-compose with no-partial-urls argument
When 0.16.3 is available

This is an alternative to https://github.com/flatpak/flatpak-builder/pull/576

We check at runtime instead of build-time and downgrade instead of
erroring. Checking at build time is insufficient as there is no
guarantee flatpak-builder is run with the same appstream version it
was compiled with.

Ideally we want to compose with `--no-partial-urls` so a warning is
printed if it is not found.

`--no-partial-urls` makes the mirrored URLs in appstream catalogue
data the full URL. So if
`--mirror-screenshot-url=https://example.org/media` and the metainfo
has
`<screenshot><image>https://github.com/foobar.png</image></screenshot>`

After composing with `--no-partial-urls`, it will become
<screenshot><image>https://example.org/media/foobar.png</image></screenshot>`
instead of `/media/foobar.png`.

This is needed by Flathub and generally should be a standard
requirement for any software store (to have a full working URL
instead of a partial whatever-that-is).