builder-git: Add a property to disable explicitly fetching LFS
The previous commit will attempt to fetch LFS by default for Git
sources even if LFS is not configured (but available).

This allows to disable that behaviour on a per source basis
since repos may have huge LFS assets which may be undesirable to fetch.

But if LFS is available and smudge filters are active, a
`git checkout` will automatically fetch LFS objects by default even
with `disable-lfs: true`. To skip that `GIT_LFS_SKIP_SMUDGE=1` can be
set in the environment.