builder-main: Always update appstream branches when exporting to repo
This creates or updates the appstream and appstream2 refs to the
exported repo. The appstream refs are needed to display appstream
information (version, name etc.) if the exported repo is being used as
a remote.

If `files/share/app-info` does not exist inside the build for example
for extensions etc. flatpak build-export simply continues by printing a
line.

Fixes https://github.com/flatpak/flatpak-builder/issues/376