builder-manifest: Override branch with --default-branch when specified
Commit 00f63cf added default-branch with the stated intention of being
able to override the branch specified in the manifest. The name is a bit
confusing.

However the scenario tested was only when there was no
specified branch at all, and thus the behavior was a bit buggy.

Till now, if you had both specified "branch": "something" in the
manifest and used --default-branch=test, the end result in the
ostree repo would be for flatpak-builder to use the branch name
specified in the manifest.

```
$ flatpak-builder --default-branch=test --force-clean --repo=_build/repo/ app/ tests/test-default-branch.json
$ ostree refs --repo=_build/repo/
app/org.test.DefaultBranch/x86_64/something
```

With this change, we properly pick up default-branch if its set,
since it can only be set through the cli and its the explicit intention
that we want to override it.

```
$ meson devenv -C _build/ flatpak-builder  --default-branch=test --user --force-clean --repo=repo_test/ app/ ../tests/test-default-branch.json
$ ostree refs --repo=_build/repo_test
app/org.test.DefaultBranch/x86_64/test
```